requirements.txt
Flask
python-telegram-bot==20.0
nest_asyncio
python-dotenv
